app-timestamp/requirements.txt
streamlit
elasticsearch
sentence-transformers
openai
psycopg2-binary
python-dotenv
sentence-transformers
ollama